separation of notifications into different handlers
